+++ packages/client/src/utils/Query.ts
@@ -2,15 +2,18 @@ function query(): string {
   return typeof window !== 'undefined' ? window.location.search.slice(1) : '';
 }
 
-function changeQuery(params: URLSearchParams): void {
+function changeQuery(params: URLSearchParams, push = false): void {
   if (typeof window === 'undefined') {
     return;
   }
   const str = params.toString();
-  if (str) {
-    window.history.replaceState({}, document.title, `${window.location.pathname}?${str}${window.location.hash}`);
+  const url = str
+    ? `${window.location.pathname}?${str}${window.location.hash}`
+    : window.location.pathname + window.location.hash;
+  if (push) {
+    window.history.pushState({}, document.title, url);
   } else {
-    window.history.replaceState({}, document.title, window.location.pathname + window.location.hash);
+    window.history.replaceState({}, document.title, url);
   }
 }
 
@@ -22,16 +25,16 @@ function get(key: string, def?: string): string | null {
   return result === null && def !== undefined ? def : result;
 }
 
-function set(key: string, value: string): void {
+function set(key: string, value: string, push = false): void {
   const params = new URLSearchParams(query());
   params.set(key, value);
-  changeQuery(params);
+  changeQuery(params, push);
 }
 
-function del(key: string): void {
+function del(key: string, push = false): void {
   const params = new URLSearchParams(query());
   params.delete(key);
-  changeQuery(params);
+  changeQuery(params, push);
 }
 
 export default { get, set, del };